Adding a simplified minimal docker image build file.
This starts with the latest, minimal Ubuntu LTS image (~27MB) and installs only the minimum requirements to run the fuzzingbook notebooks.
However, since you want to Git inside of the container, the installation of Git and its dependencies, as well as the repository clone itself almost doubles the image in size.
I understand that this easier for the students to handle instead of having an external clone that is mounted through the docker daemon, yet that would be the better solution. I’ll give this a quick test with some notebooks and add a Makefile.