ConcolicFuzzer: Use a binary tree to store explored paths
This shifts the SimpleConcolicFuzzer away from simple random
choice of predicate chains to a binary tree of decisions made
during multiple fuzzing runs. With such a data structure, the
leaves represent paths that are yet to be explored, which
allows some intelligence in choosing a reasonable path to
explore.

It fixes the problem that Anabelle faced in her project.

Also adds an initializer list so that it is easier to initialize