InformationFlow: add an initializer list
For the InformationFlow chapter, we make use of python reflection
capabilities to construct proxy methods for the string class on the
fly. Unfortunately, our code generator that converts the ipynb to
py files ignores these blocks. Hence, we need a better way to
initialize the module and generate these proxy methods when the
module is imported. The solution I have used here is to use a
`initialize()` method that need to be called after import.

Once we have a more intelligent code generator, we may be able to hook
up the initilization in the body of the module itself.