Parser: split canonical into two parts
Previously, the canonical() method took an argument `letters`, which
if true, caused longer terminal symbols to be broken into characters.
This is now split into its own procedure `single_char_tokens()`. The
utility of this approach is that, unlike `canonical()`,
`single_char_tokens()` is idempotent. This allows us to blindly apply
this procedure to all canonical grammars, and hence simplifies the
interaction between arguments, eliminating the `letters` argument.