Parser: Bug#73 Handle ambiguity better
Previously, we were using `zip` to generate combinations of items
from an array. `zip` is not the correct function to use here. It is
`itertools.product` Fixing this also exposed the problem that we did
not allow sufficient lazyness. Hence, added a lazyextractor that shows
how to deal with infinite parses.