GUIFuzzer.ipynb: Import notebook dependencies again in function
Digraph and IPython were only imported when executing the GUIFuzzer.py
directly and not if you import code from it. This was done so that code
from The Fuzzing Book can be imported even if these libraries aren't
installed. But this also prevented the import and use of functions which
use these libraries. Importing these libraries at the function level
provides both. The dependencies don't have to be installed to import
code from The Fuzzing Book. And if these functions should be used, they
import the required dependencies themselves and can be executed.

Fixes #75