Never use `python` or `pip` directly as the PATH in the Jupyter notebook may differ
Previously, we invoked `python` and `pip` which resulted in a random python and
its installation being clobbered, and resulting in assertion failures. This checkin
uses `{sys.executable}` instead. Further, it also ensures that uds-se versions of
the external repos are used. In time, we also want to make sure that we do not
clobber the home directory confirguration files such as `fuzzmanager.conf` (which is
still done presently).